refactor(plugins): scaffold core+plugins split, extract crypto module
Begin the agent-id repo split into one shared library (agent-id-core)
plus four task-scoped plugins (setup, git, vault, auth) living side by
side in plugins/. The marketplace.json gains entries for each of the
five new plugins under metadata.pluginRoot=./plugins; the existing
alien-agent-id entry is left intact for now.

First code move: pure crypto primitives (canonical JSON, b64url helpers,
Ed25519 keygen/sign/verify, JWK thumbprint, DPoP proof builder, RS256
and EdDSA JWT verifiers, SSH key conversion) relocate from
skills/alien-agent-id/lib.mjs to plugins/agent-id-core/lib/crypto.mjs.
The old lib.mjs re-exports the same surface with `export *`, so every
existing import path (CLI, tests) keeps working unchanged. No public
behavior changes. All 184 tests pass.
