+++ plugins/agent-id-vault/.claude-plugin/plugin.json
@@ -1,0 +1,9 @@
+{
+  "name": "agent-id-vault",
+  "version": "0.0.0",
+  "description": "Encrypted credential vault keyed off the agent's private key. Store, retrieve, list, and remove external-service credentials.",
+  "keywords": ["credential-vault", "agent-identity"],
+  "dependencies": [
+    { "name": "agent-id-core", "version": "^0.0.0" }
+  ]
+}
